+++ src/pages/reviews/create/components/ReviewDetailContainer/ReviewDetailContainer.tsx
@@ -2,12 +2,11 @@ import styled from '@emotion/styled';
 import { useRouter } from 'next/router';
 import React, { useCallback, useMemo } from 'react';
 import { FieldValues } from 'react-hook-form';
-import { useMutation } from 'react-query';
 import { useRecoilValue } from 'recoil';
 
 import { useGetFlavors } from '@/apis/flavors';
-import { ICreateReviewPayload, createReview } from '@/apis/review';
-import { uploadImage } from '@/apis/upload/uploadImage';
+import { ICreateReviewPayload, useCreateReviewMutation } from '@/apis/review';
+import { useUploadImageMutation } from '@/apis/upload/uploadImage';
 import BottomFloatingButtonArea from '@/components/BottomFloatingButtonArea';
 import { BOTTOM_FLOATING_BUTTON_AREA_HEIGHT } from '@/components/BottomFloatingButtonArea/BottomFloatingButtonArea';
 import Button, { ButtonCount } from '@/components/Button';
@@ -39,8 +38,8 @@ const ReviewDetailContainer: React.FC<RecordThirdStepContainerProps> = ({
   const router = useRouter();
   const reviewForm = useRecoilValue($reviewForm);
   const { data: flavors } = useGetFlavors();
-  const { mutateAsync: uploadImageMutation } = useMutation(uploadImage);
-  const { mutateAsync: createReviewMutation } = useMutation(createReview);
+  const { uploadImage } = useUploadImageMutation();
+  const { createReview } = useCreateReviewMutation();
   // const { mutateAsync: updateRecordMutation } = useMutation(updateRecord, {
   //   onSuccess: () => {
   //     router.back();
@@ -60,16 +59,16 @@ const ReviewDetailContainer: React.FC<RecordThirdStepContainerProps> = ({
 
   const handleImageUpload = useCallback(
     async (image: FormData) => {
-      const { imageUrl } = await uploadImageMutation(image);
+      const { imageUrl } = await uploadImage(image);
 
       return imageUrl;
     },
-    [uploadImageMutation],
+    [uploadImage],
   );
 
   const handleCreateSubmit = useCallback(
     (data: FieldValues) => {
-      createReviewMutation(
+      createReview(
         {
           ...reviewForm,
           ...data,
@@ -78,7 +77,7 @@ const ReviewDetailContainer: React.FC<RecordThirdStepContainerProps> = ({
         { onSuccess: (_data) => router.push(`/record/ticket/${_data.id}?type=${NEW_TYPE}`) },
       );
     },
-    [createReviewMutation, reviewForm, beer?.id, router],
+    [createReview, reviewForm, beer?.id, router],
   );
 
   // const handleUpdateSubmit = useCallback(
